+++ port_ocean/tests/core/handlers/webhook/test_processor_manager.py
@@ -4,6 +4,7 @@ from port_ocean.core.handlers.webhook.abstract_webhook_processor import (
     AbstractWebhookProcessor,
 )
 from port_ocean.core.handlers.webhook.webhook_event import (
+    EventHeaders,
     WebhookEvent,
     WebhookEventData,
     EventPayload,
@@ -140,6 +141,54 @@ class RetryableProcessor(MockWebhookProcessor):
         return True
 
 
+class MockWebhookHandlerForProcessWebhookRequest(AbstractWebhookProcessor):
+    """Concrete implementation for testing."""
+
+    def __init__(
+        self,
+        event: WebhookEvent,
+        should_fail: bool = False,
+        fail_count: int = 0,
+        max_retries: int = 3,
+    ) -> None:
+        super().__init__(event)
+        self.authenticated = False
+        self.validated = False
+        self.handled = False
+        self.should_fail = should_fail
+        self.fail_count = fail_count
+        self.current_fails = 0
+        self.error_handler_called = False
+        self.cancelled = False
+        self.max_retries = max_retries
+
+    async def authenticate(self, payload: EventPayload, headers: EventHeaders) -> bool:
+        self.authenticated = True
+        return True
+
+    async def validate_payload(self, payload: EventPayload) -> bool:
+        self.validated = True
+        return True
+
+    async def handle_event(self, payload: EventPayload) -> WebhookEventData:
+        if self.should_fail and self.current_fails < self.fail_count:
+            self.current_fails += 1
+            raise RetryableError("Temporary failure")
+        self.handled = True
+        return WebhookEventData(kind="test", data=[{}])
+
+    def filter_event_data(self, event: WebhookEvent) -> bool:
+        """Filter the event data before processing."""
+        return True
+
+    async def cancel(self) -> None:
+        self.cancelled = True
+
+    async def on_error(self, error: Exception) -> None:
+        self.error_handler_called = True
+        await super().on_error(error)
+
+
 @pytest.fixture
 def processor_manager() -> WebhookProcessorManager:
     router = APIRouter()
@@ -230,6 +279,7 @@ async def test_shutdown_ShutsDownAllTasks(
     assert len(processor_manager._webhook_processor_tasks) == 0
 
 
+@pytest.mark.asyncio
 async def test_registerProcessor_invalidHandlerRegistration_throwsError(
     processor_manager: WebhookProcessorManager,
 ) -> None:
@@ -237,6 +287,85 @@ async def test_registerProcessor_invalidHandlerRegistration_throwsError(
 
     with pytest.raises(ValueError):
         processor_manager.register_processor("/test", object)  # type: ignore
+
+
+@pytest.fixture
+def webhook_event_for_process_webhook_request() -> WebhookEvent:
+    return WebhookEvent(
+        trace_id="test-trace",
+        payload={"test": "data"},
+        headers={"content-type": "application/json"},
+    )
+
+
+@pytest.fixture
+def processor_manager_for_process_webhook_request() -> WebhookProcessorManager:
+    return WebhookProcessorManager(APIRouter(), SignalHandler())
+
+
+@pytest.fixture
+def processor(
+    webhook_event_for_process_webhook_request: WebhookEvent,
+) -> MockWebhookHandlerForProcessWebhookRequest:
+    return MockWebhookHandlerForProcessWebhookRequest(
+        webhook_event_for_process_webhook_request
+    )
+
+
+@pytest.mark.asyncio
+async def test_processWebhookRequest_successfulProcessing(
+    processor: MockWebhookHandlerForProcessWebhookRequest,
+    processor_manager_for_process_webhook_request: WebhookProcessorManager,
+) -> None:
+    """Test successful webhook processing flow."""
+    await processor_manager_for_process_webhook_request._process_webhook_request(
+        processor
+    )
+
+    assert processor.authenticated
+    assert processor.validated
+    assert processor.handled
+    assert not processor.error_handler_called
+
+
+@pytest.mark.asyncio
+async def test_processWebhookRequest_retryTwoTimesThenSuccessfulProcessing(
+    webhook_event_for_process_webhook_request: WebhookEvent,
+    processor_manager_for_process_webhook_request: WebhookProcessorManager,
+) -> None:
+    """Test retry mechanism with temporary failures."""
+    processor = MockWebhookHandlerForProcessWebhookRequest(
+        webhook_event_for_process_webhook_request, should_fail=True, fail_count=2
+    )
+
+    await processor_manager_for_process_webhook_request._process_webhook_request(
+        processor
+    )
+
+    assert processor.handled
+    assert processor.current_fails == 2
+    assert processor.retry_count == 2
+    assert processor.error_handler_called
+
+
+@pytest.mark.asyncio
+async def test_processWebhookRequest_maxRetriesExceeded_exceptionRaised(
+    webhook_event: WebhookEvent,
+    processor_manager_for_process_webhook_request: WebhookProcessorManager,
+) -> None:
+    """Test behavior when max retries are exceeded."""
+    processor = MockWebhookHandlerForProcessWebhookRequest(
+        webhook_event, should_fail=True, fail_count=2, max_retries=1
+    )
+
+    with pytest.raises(RetryableError):
+        await processor_manager_for_process_webhook_request._process_webhook_request(
+            processor
+        )
+
+    assert processor.retry_count == processor.max_retries
+    assert processor.error_handler_called
+    assert not processor.handled
 
 
 @pytest.fixture
